Minor fixes

+++ app/templates/web/routers/_entities_router.ex
@@ -7,27 +7,43 @@ defmodule <%= _.capitalize(pluralize(name)) %>Router do
 
   get "/:id" do
     <%= name %> = Repo.get(<%= _.capitalize(name) %>, id)
-    conn.resp 200, ExJSON.generate <%= name %>
+    if <%= name %> do
+      conn.resp 200, ExJSON.generate [<% _.each(attrs, function (attr) { %><%= attr.attrName %>: <%= name %>.<%= attr.attrName %>, <% }); %>id: <%= name %>.id]
+    else 
+      conn.resp 404, ""
+    end
   end
 
   post "/" do
-    json = ExJSON.parse conn.resp_body
+    conn = conn.fetch :body
+    json = ExJSON.parse conn.req_body
     <%= name %> = <%= _.capitalize(name) %>.new(
       <% var delim = ''; _.each(attrs, function (attr) { %><%= delim %><%= attr.attrName %>: json["<%= attr.attrName %>"]<% delim = ', '; }); %>
     )
-    Repo.create(<%= name %>)
-    conn.resp 201, ExJSON.generate <%= name %>
+    <%= name %> = Repo.create(<%= name %>)
+    conn.resp 201, ExJSON.generate [<% _.each(attrs, function (attr) { %><%= attr.attrName %>: <%= name %>.<%= attr.attrName %>, <% }); %>id: <%= name %>.id]
   end
 
   put "/:id" do
-    json = ExJSON.parse conn.resp_body
-    <%= name %> = Repo.get(<%= _.capitalize(name) %>, id).update(json)
-    conn.resp 200, ExJSON.generate <%= name %>
+    conn = conn.fetch :body
+    json = ExJSON.parse conn.req_body
+    <%= name %> = Repo.get(<%= _.capitalize(name) %>, id)
+    if <%= name %> do
+      <%= name %> = <%= name %>.update(json)
+      Repo.update(<%= name %>)
+      conn.resp 200, ExJSON.generate [<% _.each(attrs, function (attr) { %><%= attr.attrName %>: <%= name %>.<%= attr.attrName %>, <% }); %>id: <%= name %>.id]
+    else 
+      conn.resp 404, ""
+    end
   end
 
   delete "/:id" do
     <%= name %> = Repo.get(<%= _.capitalize(name) %>, id)
-    Repo.delete(<%= name %>)
-    conn.resp 200, ""
+    if <%= name %> do
+      Repo.delete(<%= name %>)
+      conn.resp 200, ""
+    else 
+      conn.resp 404, ""
+    end
   end
 end
